add second enemy and his logic
