+++ UnityProject/Assets/Script/EnemyMovementTree.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: dd81e0e1d8995924dacc4424c8f49141
+timeCreated: 1490980275
+licenseType: Store
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
