+++ UnityProject/Assets/Script/EnemyMovementZombie.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c20c1c2b460bab544ad503ee5b049c8e
+timeCreated: 1490980275
+licenseType: Store
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
