+++ UnityProject/Assets/Standard Assets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2f4f15bd6d2ef784bb15152940668f2c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
