+++ UnityProject/Assets/Standard Assets/Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b88065e2898bd44b8915287a2784f56
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
