+++ UnityProject/Assets/Standard Assets/Effects/ImageEffects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b70a60d4f20f654ca069b920b0bae92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
