+++ UnityProject/Assets/Standard Assets/Effects/ImageEffects/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60426c184c4e53b4e8e58a802ebcfaad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
