+++ UnityProject/Assets/_GhoulZombie.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2d385018161a8c046bce4f987235abdb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
