+++ UnityProject/Assets/_GhoulZombie/Black.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 5cd795b281bbe8c4b9a49869782d03ee
+timeCreated: 1521472153
+licenseType: Store
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
