+++ UnityProject/Assets/_GhoulZombie/Ghoul.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 227ee07e9e5b7914ba8fbb8f430be9a5
+timeCreated: 1521477726
+licenseType: Store
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
