+++ UnityProject/Assets/_GhoulZombie/GhoulMaterial.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 7e013ff2514571042a1ac03d3b6f9eef
+timeCreated: 1521470644
+licenseType: Store
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
