+++ UnityProject/Assets/_GhoulZombie/GhoulScene.unity.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 85b08af126c9990409c1c4ed5eec814d
+timeCreated: 1521471397
+licenseType: Store
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
