+++ UnityProject/Assets/_GhoulZombie/GhoulSceneSettings.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 00d93666b36044848b80358a6278c143
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
