+++ UnityProject/Assets/_GhoulZombie/PPS_Ghoul.asset.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 01ea3b62903a0b4449c9b786c0ca6b8d
+timeCreated: 1513905989
+licenseType: Store
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
